requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
2captcha-python>=1.2.0
google-generativeai>=0.3.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
